+++ src/staff/index.ts
@@ -1,0 +1,2 @@
+export { default as VentanasSection } from './VentanasSection'
+export type { VentanasSectionProps } from './VentanasSection'